Move webview communication to its own folder
