+++ apps/vs-code-react/src/run-service/types.ts
@@ -1,7 +1,3 @@
-import type { InitializePayload, Status } from '../state/WorkflowSlice';
-import type { ApiHubServiceDetails, SchemaType } from '@microsoft/logic-apps-shared';
-import type { MapDefinitionData, ExtensionCommand, ConnectionsData, IDesignerPanelMetadata } from '@microsoft/vscode-extension';
-
 export interface IApiService {
   getWorkflows(subscriptionId: string, iseId?: string, location?: string): Promise<WorkflowsList[]>;
   getSubscriptions(): Promise<Array<ISubscription>>;
@@ -200,101 +196,6 @@ export const StyledWorkflowPart = {
   workflow: 'Workflow',
 };
 export type StyledWorkflowPart = (typeof StyledWorkflowPart)[keyof typeof StyledWorkflowPart];
-
-type FetchSchemaData = { fileName: string; type: SchemaType };
-export type XsltData = { filename: string; fileContents: string };
-
-// Data Mapper Message Interfaces
-export interface FetchSchemaMessage {
-  command: typeof ExtensionCommand.fetchSchema;
-  data: FetchSchemaData;
-}
-
-export interface LoadDataMapMessage {
-  command: typeof ExtensionCommand.loadDataMap;
-  data: MapDefinitionData;
-}
-
-export interface ShowAvailableSchemasMessage {
-  command: typeof ExtensionCommand.showAvailableSchemas;
-  data: string[];
-}
-
-export interface GetAvailableCustomXsltPathsMessage {
-  command: typeof ExtensionCommand.getAvailableCustomXsltPaths;
-  data: string[];
-}
-
-export interface SetXsltDataMessage {
-  command: typeof ExtensionCommand.setXsltData;
-  data: XsltData;
-}
-
-export interface SetRuntimePortMessage {
-  command: typeof ExtensionCommand.setRuntimePort;
-  data: string;
-}
-
-export interface GetConfigurationSettingMessage {
-  command: typeof ExtensionCommand.getConfigurationSetting;
-  data: boolean;
-}
-
-// Designer Message Interfaces
-export interface ReceiveCallbackMessage {
-  command: typeof ExtensionCommand.receiveCallback;
-  data: any;
-}
-
-export interface CompleteFileSystemConnectionMessage {
-  command: typeof ExtensionCommand.completeFileSystemConnection;
-  data: { connectionName: string; connection: any; error: string };
-}
-
-export interface UpdatePanelMetadataMessage {
-  command: typeof ExtensionCommand.update_panel_metadata;
-  data: {
-    panelMetadata: IDesignerPanelMetadata;
-    connectionData: ConnectionsData;
-    apiHubServiceDetails: ApiHubServiceDetails;
-  };
-}
-
-// Rest of Message Interfaces
-export interface InjectValuesMessage {
-  command: typeof ExtensionCommand.initialize_frame;
-  data: InitializePayload & {
-    project: string;
-  };
-}
-
-export interface UpdateAccessTokenMessage {
-  command: typeof ExtensionCommand.update_access_token;
-  data: {
-    accessToken?: string;
-  };
-}
-
-export interface UpdateExportPathMessage {
-  command: typeof ExtensionCommand.update_export_path;
-  data: {
-    targetDirectory: ITargetDirectory;
-  };
-}
-
-export interface AddStatusMessage {
-  command: typeof ExtensionCommand.add_status;
-  data: {
-    status: string;
-  };
-}
-
-export interface SetFinalStatusMessage {
-  command: typeof ExtensionCommand.set_final_status;
-  data: {
-    status: Status;
-  };
-}
 
 export interface IExportDetails {
   exportDetailCategory: string;
